refactor: move virtualizer to component-base
