+++ packages/combo-box/package.json
@@ -40,8 +40,7 @@
     "@vaadin/vaadin-lumo-styles": "22.0.0-alpha10",
     "@vaadin/vaadin-material-styles": "22.0.0-alpha10",
     "@vaadin/vaadin-overlay": "22.0.0-alpha10",
-    "@vaadin/vaadin-themable-mixin": "22.0.0-alpha10",
-    "@vaadin/virtual-list": "22.0.0-alpha10"
+    "@vaadin/vaadin-themable-mixin": "22.0.0-alpha10"
   },
   "devDependencies": {
     "@esm-bundle/chai": "^4.3.4",
