+++ packages/component-base/test/virtualizer.test.js
